migrated the game function to menu instead of controller

+++ src/main/java/com/example/cs102/Comparators/SuitComparator.java
@@ -5,7 +5,7 @@ import com.example.cs102.poker.Card;
 public class SuitComparator implements Comparator <Card>{
     public int compare (Card card1 , Card card2){
         if (card1.getSuit() != card2.getSuit()){
-            return (int)(card1.getSuit() - card2.getSuit());
+            return (int)(card1.getSuit() - (int)card2.getSuit());
         }
         return card1.getValue() - card2.getValue();
     }
